module-avb: fixing banner header, and splitting commands
