+++ src/modules/meson.build
@@ -750,7 +750,10 @@ if build_module_avb
       'module-avb/srp.c',
       'module-avb/stream.c',
       'module-avb/aecp-cmd-resp/aecp-aem-available.c',
-      'module-avb/aecp-cmd-resp/aecp-aem-lock.c'
+      'module-avb/aecp-cmd-resp/aecp-aem-descriptors.c',
+      'module-avb/aecp-cmd-resp/aecp-aem-get-avb-info.c',
+      'module-avb/aecp-cmd-resp/aecp-aem-lock.c',
+      'module-avb/aecp-cmd-resp/aecp-aem-unsol-notifications.c'
       ],
     include_directories : [configinc],
     install : true,
