+++ src/modules/module-avb/aecp-aem-state.h
@@ -1,5 +1,6 @@
 /* AVB support */
-/* SPDX-FileCopyrightText: Copyright © 2025 Alexandre Malki */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alexandre Malki <alexandre.malki@kebag-logic.com> */
 /* SPDX-License-Identifier: MIT */
 
 #ifndef AVB_AECP_AEM_STATE_H
