+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-cmd-resp-common.h
@@ -1,0 +1,12 @@
+/* AVB support */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alex Malki <alexandre.malki@kebag-logic.com> */
+/* SPDX-License-Identifier: MIT  */
+
+#ifndef __AVB_AECP_AEM_CMD_RESP_COMMON_H__
+#define __AVB_AECP_AEM_CMD_RESP_COMMON_H__
+
+#include <stdint.h>
+#include "../aecp-aem.h"
+
+#endif // __AVB_AECP_AEM_AVAILABLE_H__
