+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-lock.c
@@ -1,6 +1,9 @@
+/* AVB support */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alex Malki <alexandre.malki@kebag-logic.com> */
+/* SPDX-License-Identifier: MIT  */
 
 #include <limits.h>
-
 
 #include "../aecp-aem-state.h"
 #include "../descriptors.h"
