+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-lock.h
@@ -1,10 +1,15 @@
+/* AVB support */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alex Malki <alexandre.malki@kebag-logic.com> */
+/* SPDX-License-Identifier: MIT  */
+
 #ifndef __AVB_AECP_AEM_LOCK_H__
 #define __AVB_AECP_AEM_LOCK_H__
 
 #define AECP_AEM_LOCK_ENTITY_EXPIRE_TIMEOUT (60UL)
 #define AECP_AEM_LOCK_ENTITY_FLAG_LOCK		(1)
 
-#include "../aecp-aem.h"
+#include "aecp-aem-cmd-resp-common.h"
 
 /**
  * @brief Command handling will generate the response for the lock command
