+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-types.h
@@ -1,3 +1,8 @@
+/* AVB support */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alex Malki <alexandre.malki@kebag-logic.com> */
+/* SPDX-License-Identifier: MIT  */
+
 #ifndef __AVB_AECP_AEM_TYPES_H__
 #define __AVB_AECP_AEM_TYPES_H__
