+++ src/modules/module-avb/aecp-state-vars.h
@@ -1,3 +1,8 @@
+/* AVB support */
+/* SPDX-FileCopyrightText: Copyright © 2025 Kebag-Logic */
+/* SPDX-FileCopyrightText: Copyright © 2025 Alexandre Malki <alexandre.malki@kebag-logic.com> */
+/* SPDX-License-Identifier: MIT */
+
 #ifndef  AECP_STATE_VARS_H_
 #define  AECP_STATE_VARS_H_
